fix missing files in previous commit

+++ src/main/java/fr/pollux/katarpn/Operation.java
@@ -7,7 +7,7 @@ public class Operation
         return num1 + num2;
     }
 
-    public static int substraction(int num1, int num2)
+    public static double substraction(double num1, double num2)
     {
         return num2 - num1;
     }
