+++ src/main/java/fr/pollux/katarpn/RPNHandler.java
@@ -4,20 +4,19 @@ import fr.pollux.katarpn.exception.InvalidFormatException;
 import fr.pollux.katarpn.Operation;
 
 import java.util.*;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
 public class RPNHandler {
     private List<String> tokens;
-    private Stack<Integer> stack;
+    private Stack<Double> stack;
 
     public RPNHandler(String rpn) {
         tokens = Arrays.asList(rpn.split(" "));
         stack = new Stack<>();
     }
 
-     public int operations() throws InvalidFormatException {
+     public double operations() throws InvalidFormatException {
         try {
             for (String el : tokens) {
                 switch (el) {
@@ -27,6 +26,7 @@ public class RPNHandler {
                     case "-":
                         stack.add(Operation.substraction(stack.pop(), stack.pop()));
                         break;
+                    
                     default:
                         stack.add(Double.parseDouble(el));
                 }
